+++ src/euroeval/dataset_configs/lithuanian.py
@@ -47,7 +47,6 @@ LT_HISTORY_CONFIG = DatasetConfig(
     huggingface_id="EuroEval/lt-history",
     task=KNOW,
     languages=[LT],
-    splits=["train", "test"],
 )
 
 WINOGRANDE_LT_CONFIG = DatasetConfig(
@@ -57,6 +56,5 @@ WINOGRANDE_LT_CONFIG = DatasetConfig(
     huggingface_id="EuroEval/winogrande-lt",
     task=COMMON_SENSE,
     languages=[LT],
-    splits=["train", "test"],
     _labels=["a", "b"],
 )